ftlassignment/requirements.txt
asgiref==3.2.7
boto3==1.13.24
botocore==1.16.24
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-appconf==1.0.4
django-filter==2.2.0
django-qr-code==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
httpie==2.1.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
psycopg2==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
qrcode==6.1
requests==2.23.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
